pyproject.toml
[project]
name = "hedra-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-anam>=1.2.11",
    "livekit-plugins-hedra>=1.2.11",
    "livekit-plugins-noise-cancellation~=0.2",
    "pillow>=11.3.0",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
]
